90cent/requirements.txt
requests>=2.31.0
websocket-client>=1.6.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
ta-lib>=0.4.28
aiohttp>=3.9.0
python-dateutil>=2.8.2
py-clob-client>=0.28.0
py-order-utils>=0.1.0
flask>=3.0.0
polymarket-apis>=0.4.0
httpx[http2]>=0.27.0
web3>=7.0.0
